pyproject.toml
[project]
name = "email-triage-env"
version = "1.0.0"
description = "A real-world customer support email triage environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "openai"
]

[project.scripts]
server = "server.app:main"
